requirements.txt
flask
flask-cors
google-genai
python-dotenv
firebase-admin
gunicorn